feat: seperate parser
