+++ lib/flutter_cache.dart
@@ -2,6 +2,7 @@ library flutter_cache;
 
 import 'dart:convert';
 import 'package:shared_preferences/shared_preferences.dart';
+import 'package:flutter_cache/Parse.dart';
 
 class Cache {
 
@@ -165,41 +166,5 @@ class Cache {
   static int _currentTimeInSeconds() { // private
     var ms = (new DateTime.now()).millisecondsSinceEpoch;
     return (ms / 1000).round();
-  }
-}
-
-class Parse {
-
-  static content (var data) {
-    /* @type String */
-    if (data is String) return {
-      'content' : data,
-      'type'    : 'String'
-    };
-
-    /* @type Map */
-    if (data is Map) return {
-      'content' : jsonEncode(data),
-      'type'    : 'Map'
-    };
-
-    /* @type List<String> */
-    if (data is List<String>) {
-      return {
-        'content' : data,
-        'type'    : 'List<String>'
-      };
-    }
-
-    /* @type List<Map> */
-    if (data is List<Map>) {
-      List<String> list = data.map((i) => jsonEncode(i)).toList();
-      return {
-        'content' : list,
-        'type'    : 'List<Map>'
-      };
-    }
-
-    throw ('Unsupported Data Type');
   }
 }
